docker-migration-tool/requirements.txt
docker
docker-compose
requests
paramiko
pyyaml
zipfile36